+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * [#913](https://github.com/clojure-emacs/cider-nrepl/pull/913): Disable background warmup of `orchard.java` cache.
 * [#913](https://github.com/clojure-emacs/cider-nrepl/pull/913): Enable background warmup of Compliment cache.
-* TBD-PR: Sort printed maps in test output
+* [#917](https://github.com/clojure-emacs/cider-nrepl/pull/917): Sort printed maps in test output
 
 ## 0.52.1 (2025-02-24)